fix hideElement(installBtnComponent)
